requirements.txt
attrs==21.2.0
autopep8==1.5.7
click==8.0.1
dnspython==1.16.0
Flask==2.0.1
Flask-Cors==3.0.10
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
ply==3.11
pycodestyle==2.7.0
pymongo==3.12.0
pyrsistent==0.18.0
PyYAML==5.4.1
six==1.16.0
toml==0.10.2
Werkzeug==2.0.1
